Cancel processing

+++ src/main/java/lt/luminor/payments/repository/PaymentRepository.java
@@ -1,5 +1,7 @@
 package lt.luminor.payments.repository;
 
+import java.math.BigDecimal;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -13,6 +15,6 @@ import lt.luminor.payments.entity.PaymentStatus;
 public interface PaymentRepository extends JpaRepository<Payment, Long> {
 	Page<Payment> findByCreatedByAndPaymentStatusOrderByPaymentType(Long createdBy, PaymentStatus paymentStatus, Pageable pageable);
 	Page<Payment> findByCreatedByAndCurrencyAndPaymentStatusOrderByAmount(Long createdBy, Currency currency, PaymentStatus paymentStatus, Pageable pageable);
-	Page<Payment> findByCreatedByAndCurrencyAndPaymentStatusAndAmountLessThanOrderByAmount(Long createdBy, Currency currency, PaymentStatus paymentStatus, Double amount, Pageable pageable);
-	Page<Payment> findByCreatedByAndCurrencyAndPaymentStatusAndAmountGreaterThanOrderByAmount(Long createdBy, Currency currency, PaymentStatus paymentStatus, Double amount, Pageable pageable);
+	Page<Payment> findByCreatedByAndCurrencyAndPaymentStatusAndAmountLessThanOrderByAmount(Long createdBy, Currency currency, PaymentStatus paymentStatus, BigDecimal amount, Pageable pageable);
+	Page<Payment> findByCreatedByAndCurrencyAndPaymentStatusAndAmountGreaterThanOrderByAmount(Long createdBy, Currency currency, PaymentStatus paymentStatus, BigDecimal amount, Pageable pageable);
 }
